# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: GrowthForge AI API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects/{id}/campaigns:
    get:
      operationId: listCampaigns
      tags: [campaigns]
      summary: List ad campaigns
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of campaigns
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Campaign"
    post:
      operationId: createCampaign
      tags: [campaigns]
      summary: Create/generate an ad campaign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CampaignInput"
      responses:
        "201":
          description: Created campaign
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
  /projects/{id}/campaigns/{campaignId}:
    get:
      operationId: getCampaign
      tags: [campaigns]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: campaignId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Campaign
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
    patch:
      operationId: updateCampaign
      tags: [campaigns]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: campaignId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CampaignUpdate"
      responses:
        "200":
          description: Updated campaign
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
  /projects/{id}/campaigns/{campaignId}/performance:
    get:
      operationId: getCampaignPerformance
      tags: [campaigns]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: campaignId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Campaign performance metrics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CampaignPerformance"

  # ── Assets ─────────────────────────────────────────────────────────────────
  /projects/{id}/assets:
    get:
      operationId: listAssets
      tags: [assets]
      summary: List all assets in library
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of assets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Asset"

  # ── Analytics ──────────────────────────────────────────────────────────────
  /projects/{id}/analytics:
    get:
      operationId: getProjectAnalytics
      tags: [analytics]
      summary: Get project analytics overview
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Analytics data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectAnalytics"
  /projects/{id}/reports:
    get:
      operationId: listReports
      tags: [analytics]
      summary: List generated reports
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of reports
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Report"
    post:
      operationId: generateReport
      tags: [analytics]
      summary: Generate a report
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportInput"
      responses:
        "200":
          description: Generated report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"

  # ── AI Agent ───────────────────────────────────────────────────────────────
  /projects/{id}/agent/chat:
    post:
      operationId: agentChat
      tags: [agent]
      summary: Send a message to the AI marketing agent
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentMessageInput"
      responses:
        "200":
          description: Agent response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentMessage"
  /projects/{id}/agent/history:
    get:
      operationId: getAgentHistory
      tags: [agent]
      summary: Get AI agent conversation history
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Conversation history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AgentMessage"
components:
  schemas:
...
    ContentAsset:
      type: object
      required: [id, projectId, type, title, status, createdAt]
      properties:
        id:
          type: integer
        projectId:
          type: integer
        type:
          type: string
        title:
          type: string
        body:
          type: ["string", "null"]
        metaDescription:
          type: ["string", "null"]
        seoKeywords:
          type: ["string", "null"]
        status:
          type: string
        hookStrength:
          type: ["integer", "null"]
        conversionPotential:
          type: ["integer", "null"]
        engagementPotential:
          type: ["integer", "null"]
        viralPotential:
          type: ["integer", "null"]
        createdAt:
          type: string

    # ── Social ───────────────────────────────────────────────────────────────
...
    Video:
      type: object
      required: [id, projectId, title, type, status, createdAt]
      properties:
        id:
          type: integer
        projectId:
          type: integer
        title:
          type: string
        type:
          type: string
        status:
          type: string
        script:
          type: ["string", "null"]
        storyboard:
          type: ["string", "null"]
        voiceover:
          type: ["string", "null"]
        thumbnailUrl:
          type: ["string", "null"]
        videoUrl:
          type: ["string", "null"]
        duration:
          type: ["integer", "null"]
        hookStrength:
          type: ["integer", "null"]
        engagementPotential:
          type: ["integer", "null"]
        viralPotential:
          type: ["integer", "null"]
        createdAt:
          type: string
...
    CampaignInput:
      type: object
      required: [name, platform]
      properties:
        name:
          type: string
        platform:
          type: string
        budget:
          type: number
        prompt:
          type: string
    CampaignUpdate:
      type: object
      properties:
        name:
          type: string
        status:
          type: string
        budget:
          type: number
        autonomousOptimization:
          type: boolean
    Campaign:
      type: object
      required: [id, projectId, name, platform, status, createdAt]
      properties:
        id:
          type: integer
        projectId:
          type: integer
        name:
          type: string
        platform:
          type: string
        status:
          type: string
        budget:
          type: ["number", "null"]
        spent:
          type: ["number", "null"]
        impressions:
          type: ["integer", "null"]
        clicks:
          type: ["integer", "null"]
        conversions:
          type: ["integer", "null"]
        roas:
          type: ["number", "null"]
        ctr:
          type: ["number", "null"]
        cpc:
          type: ["number", "null"]
        autonomousOptimization:
          type: boolean
        createdAt:
          type: string
    CampaignPerformance:
      type: object
      required: [campaignId, period]
      properties:
        campaignId:
          type: integer
        period:
          type: string
        impressions:
          type: integer
        clicks:
          type: integer
        conversions:
          type: integer
        spend:
          type: number
        revenue:
          type: number
        roas:
          type: number
        ctr:
          type: number
        cpc:
          type: number
        cpa:
          type: number

    # ── Assets ───────────────────────────────────────────────────────────────
    Asset:
      type: object
      required: [id, projectId, type, name, createdAt]
      properties:
        id:
          type: integer
        projectId:
          type: integer
        type:
          type: string
        name:
          type: string
        url:
          type: ["string", "null"]
        thumbnailUrl:
          type: ["string", "null"]
        fileSize:
          type: ["integer", "null"]
        mimeType:
          type: ["string", "null"]
        createdAt:
          type: string

    # ── Analytics ────────────────────────────────────────────────────────────
    ProjectAnalytics:
      type: object
      required: [projectId, period]
      properties:
        projectId:
          type: integer
        period:
          type: string
        websiteTraffic:
          type: integer
        leads:
          type: integer
        revenue:
          type: number
        adSpend:
          type: number
        roas:
          type: number
        topContent:
          type: array
          items:
            $ref: "#/components/schemas/ContentAsset"
        topVideos:
          type: array
          items:
            $ref: "#/components/schemas/Video"
        chartData:
          type: array
          items:
            $ref: "#/components/schemas/ChartDataPoint"
    ChartDataPoint:
      type: object
      required: [date, value]
      properties:
        date:
          type: string
        value:
          type: number
        label:
          type: string
    ReportInput:
      type: object
      required: [type]
      properties:
        type:
          type: string
        period:
          type: string
    Report:
      type: object
      required: [id, projectId, type, period, status, createdAt]
      properties:
        id:
          type: integer
        projectId:
          type: integer
        type:
          type: string
        period:
          type: string
        status:
          type: string
        summary:
          type: ["string", "null"]
        kpiTrends:
          type: ["string", "null"]
        recommendations:
          type: ["string", "null"]
        createdAt:
          type: string

    # ── Agent ────────────────────────────────────────────────────────────────
    AgentMessageInput:
      type: object
      required: [message]
      properties:
        message:
          type: string
    AgentMessage:
      type: object
      required: [id, projectId, role, content, createdAt]
      properties:
        id:
          type: integer
        projectId:
          type: integer
        role:
          type: string
        content:
          type: string
        actionType:
          type: ["string", "null"]
        actionResult:
          type: ["string", "null"]
        createdAt:
          type: string