# lib/api-zod/src/generated/types/asset.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * GrowthForge AI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Asset {
  id: number;
  projectId: number;
  type: string;
  name: string;
  /** @nullable */
  url?: string | null;
  /** @nullable */
  thumbnailUrl?: string | null;
  /** @nullable */
  fileSize?: number | null;
  /** @nullable */
  mimeType?: string | null;
  createdAt: string;
}