# lib/api-zod/src/generated/types/campaignPerformance.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * GrowthForge AI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CampaignPerformance {
  campaignId: number;
  period: string;
  impressions?: number;
  clicks?: number;
  conversions?: number;
  spend?: number;
  revenue?: number;
  roas?: number;
  ctr?: number;
  cpc?: number;
  cpa?: number;
}