# lib/api-zod/src/generated/types/video.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * GrowthForge AI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Video {
  id: number;
  projectId: number;
  title: string;
  type: string;
  status: string;
  /** @nullable */
  script?: string | null;
  /** @nullable */
  storyboard?: string | null;
  /** @nullable */
  voiceover?: string | null;
  /** @nullable */
  thumbnailUrl?: string | null;
  /** @nullable */
  videoUrl?: string | null;
  /** @nullable */
  duration?: number | null;
  /** @nullable */
  hookStrength?: number | null;
  /** @nullable */
  engagementPotential?: number | null;
  /** @nullable */
  viralPotential?: number | null;
  createdAt: string;
}